Simpler mountain program
It might be beneficial to simply turn immediatley toward the mountain
rather than going to the back wall. Though we would be on the
wrong-color mountain, this is allowed in Autonomous as long as you get
off. It would help in terms of debris and we might not need to use a
roll angle.